requirements.txt
Django
Pillow
django-widget-tweaks
geopy
icalendar
timezonefinder
weasyprint
openpyxl